AI_programs/Language_Models/WebAPI/LLM_helper_backend/LLM_helper_backend/requirements.txt
uvicorn
fastapi
python-dotenv
pydantic_settings
openai